Added status messages to dotfile processing.

+++ bootstrap.dotfiles.bash
@@ -1,5 +1,43 @@
 #!/usr/bin/env bash
 
+DEBUG='false'
+
+debug() {
+  if [ $DEBUG == 'true' ]; then
+    echo "DEBUG: $1"
+  fi
+}
+
+info() {
+    cat <<-HEADER
+
+####
+
+HEADER
+    echo "$1"
+    cat <<-FOOTER
+
+####
+
+FOOTER
+
+}
+
+error() {
+        cat <<-HEADER
+
+####
+
+HEADER
+    echo "$1"
+    cat <<-FOOTER
+
+####
+
+FOOTER
+    exit 1
+}
+
 # Presence of this file indicates whether this is a personal system
 BOOTSTRAP_PERSONAL="${HOME}/.bootstrap_personal"
 
@@ -7,7 +45,6 @@ export STARTUP_DIR="${HOME}/.startup"
 STARTUP_BACKUP_DIR="${HOME}/.startup_backup"
 
 function link_startup_file() {
-
     backup_timestamp=$(date +%Y%m%d_%H%M%S)
     startup_backup_dir="${STARTUP_BACKUP_DIR}/${backup_timestamp}"
     mkdir -p ${startup_backup_dir}
@@ -27,6 +64,7 @@ function link_startup_file() {
         fi
     fi
     target_file_name="$HOME/.${source_file_name}"
+    info "link_startup_file ${full_source_file_name} ${target_file_name}"
     if [ -f ${full_source_file_name} ]; then
         if [ -f ${target_file_name} ]; then
             cp ${target_file_name} ${startup_backup_dir}
@@ -46,14 +84,23 @@ function create_base_dotfiles_if_needed() {
     host_startup_dir=$(hostname)
     pushd ${dotfiles_dir}
     if ! [ -d ${host_startup_dir} ]; then
+        info "create_base_dotfiles_if_needed ${host_startup_dir}"
         reference_dir="$(pwd)/reference"
         mkdir -p ${host_startup_dir}
         pushd ${host_startup_dir}
         current_dir=$(pwd)
-        ln -s "$(reference_dir)/bash_aliases" "${current_dir}/bash_aliases"
-        ln -s "$(reference_dir)/bash_env" "${current_dir}/bash_env"
-        ln -s "$(reference_dir)/bash_profile" "${current_dir}/bash_profile"
-        ln -s "$(reference_dir)/bashrc" "${current_dir}/bashrc"
+        if ! [ -f "${current_dir}/bash_aliases" ]; then
+            ln -s "$(reference_dir)/bash_aliases" "${current_dir}/bash_aliases"
+        fi
+        if ! [ -f "${current_dir}/bash_env" ]; then
+            ln -s "$(reference_dir)/bash_env" "${current_dir}/bash_env"
+        fi
+        if ! [ -f "${current_dir}/bash_profile" ]; then
+            ln -s "$(reference_dir)/bash_profile" "${current_dir}/bash_profile"
+        fi
+        if ! [ -f "${current_dir}/bashrc" ]; then
+            ln -s "$(reference_dir)/bashrc" "${current_dir}/bashrc"
+        fi
         touch bash_local
         popd
     fi
@@ -64,7 +111,7 @@ mkdir -p ${STARTUP_DIR}
 mkdir -p ${STARTUP_BACKUP_DIR}
 
 cd ${STARTUP_DIR}
-
+å
 if ! [ -d bash-it ]; then
     git clone https://github.com/Bash-it/bash-it.git
 else
@@ -100,6 +147,7 @@ if [ -f ${BOOTSTRAP_PERSONAL} ]; then
     fi
 
     # Link startup files locally to host-based directory (as for today)
+    info "Linking startup files"
     source_prefix=$(hostname)
     link_startup_file "bash_env" "$source_prefix"
     link_startup_file "bashrc" "$source_prefix"
@@ -132,6 +180,8 @@ else
         git pull origin master
         popd
     fi
+
+    info "Linking local startup files"
     # Link startup files locally
     link_startup_file "bash_env"
     link_startup_file "bashrc"
@@ -140,7 +190,7 @@ else
     link_startup_file "bash_profile"
 fi
 
-# Link MYBIN to HOME/binp
+# Link MYBIN to HOME/bin
 
 function link_bin_file() {
     source_file=$1
@@ -150,11 +200,13 @@ function link_bin_file() {
         rm -f ${full_target_file}
     fi
     if ! [ -e ${full_target_file} ]; then
+        info "link_bin_file ${full_source_file} ${full_target_file}"
         ln -s ${full_source_file} ${full_target_file}
     fi
 }
 
 function link_bin_files() {
+    info "link_bin_files"
     for bin_file in $(ls -1 "$STARTUP_DIR/mybin")
     do
         link_bin_file ${bin_file}
@@ -165,10 +217,9 @@ mkdir -p "$HOME/bin"
 link_bin_files
 
 # Enable useful aliases
+info "Enabling useful aliases"
 source ${HOME}/.bash_profile
 bash-it enable alias general git
 bash-it enable plugin alias-completion base
 my-bash-it enable alias general git osx python vagrant
 my-bash-it enable plugin general git python sublime ttitle
-
-
